Replaced SQLite with PersistentStore.
All SQLite removed and replaced Vectors which are stored in the
PersistentStore.

The activity messages are simply stored as Strings in an activity log
Vector.
For the list of files, the FileHolder was refactored to make it
Persistable (by implementing the empty interface). Each file has a
corresponding FileHolder object which is stored in a file log Vector.