New, and better, Logging system.
More informative: Supports different levels: warn, info, etc.

More useful: Logs to the EventLogger if the DEBUG flag is set.

More concise: A logging event doesn't require a tag to be specified;
the tag is specified by the class when it requests a logger. Internally,
each class has its own logger which is stored in a hashmap as
<class>=><logger instance>.